update tests to use shared fixtures
