+++ tests/conftest.py
@@ -1,10 +1,31 @@
 import pytest
-from tests.fixtures.fixture_artifact import sample_fixture
-from unittest.mock import patch, MagicMock
+from tests.fixtures.fixture_artifact import client
+from unittest.mock import patch, MagicMock, Mock
+
 
 @pytest.fixture
 def mock_db_connection():
-    with patch('utils.db_config.get_db_connection') as mock_get_db_connection:
+    with patch('app.get_db_connection') as mock_get_db_connection:
         mock_conn = MagicMock()
         mock_get_db_connection.return_value.__enter__.return_value = mock_conn
         yield mock_conn
+
+
+@pytest.fixture
+def mock_dependencies():
+    with patch('app.recursive_repo_clone') as mock_clone, \
+            patch('app.build_docker_image') as mock_docker_build, \
+            patch('app.start_ledger') as mock_start_ledger, \
+            patch('app.open', create=True) as mock_open:
+
+        mock_docker = Mock()
+        mock_docker.save.return_value = [b'mock_image_data']
+        mock_docker_build.return_value = mock_docker
+        mock_start_ledger.return_value = ({"status": "success"}, 200)
+
+        yield {
+            'clone': mock_clone,
+            'docker_build': mock_docker_build,
+            'start_ledger': mock_start_ledger,
+            'open': mock_open
+        }
