+++ tests/fixtures/fixture_artifact.py
@@ -1,5 +1,6 @@
-import pytest 
+import pytest
 from app import app
+
 
 @pytest.fixture
 def client():
@@ -7,5 +8,3 @@ def client():
     app.config['TESTING'] = True
     with app.test_client() as client:
         yield client
-
-
