+++ tests/test_functions/test_create_ledger.py
@@ -1,41 +1,8 @@
 import pytest
 from app import app
 from unittest.mock import patch, Mock
-
-
-@pytest.fixture
-def client():
-    app.config['TESTING'] = True
-    with app.test_client() as client:
-        yield client
-
-
-@pytest.fixture
-def mock_db_connection():
-    with patch('app.get_db_connection') as mock_conn:
-        mock_context = Mock()
-        mock_conn.return_value.__enter__.return_value = mock_context
-        yield mock_context
-
-
-@pytest.fixture
-def mock_dependencies():
-    with patch('app.recursive_repo_clone') as mock_clone, \
-            patch('app.build_docker_image') as mock_docker_build, \
-            patch('app.start_ledger') as mock_start_ledger, \
-            patch('app.open', create=True) as mock_open:
-
-        mock_docker = Mock()
-        mock_docker.save.return_value = [b'mock_image_data']
-        mock_docker_build.return_value = mock_docker
-        mock_start_ledger.return_value = ({"status": "success"}, 200)
-
-        yield {
-            'clone': mock_clone,
-            'docker_build': mock_docker_build,
-            'start_ledger': mock_start_ledger,
-            'open': mock_open
-        }
+from tests.fixtures.fixture_artifact import client
+from tests.conftest import mock_db_connection, mock_dependencies
 
 
 def test_create_ledger_success(client, mock_db_connection, mock_dependencies):
